#35 fix reindex so it picks up recent updated records
In run.py/reindex, when force processing flag (-o) is not set and a record has not been sent to solr (processed field is null) it is never sent to Solr.  When processed is null utils get_date returns now.  It looks like the record has just been updated so it is not queued for writing to solr.  The change follows code in tasks.py/task_index_records, if the processed time is null, it defaults to 1972.